SwiftUI Todo List: User Experience & Animations

+++ Swiftful Thinking SwiftUI Bootcamp/SwiftulThinkingSwiftUIToDoList/SwiftulThinkingSwiftUIToDoList/Views/ListView.swift
@@ -13,23 +13,31 @@ struct ListView: View {
     
     var body: some View {
         // ItemModel은 Identifiable protocol을 채택하고 있으므로 id를 정해주지 않아도 된다.
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                /* NoItemView가 나타날 때 transition 효과 */
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {    
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    /* onDelete는 List에 직접 붙일 수 없음
+                     * ForEach에 붙여야 한다.
+                     * onDelete를 붙여야 edit button을 누를 때 삭제 버튼이 나타난다.
+                     */
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(.plain)
             }
-            /* onDelete는 List에 직접 붙일 수 없음
-             * ForEach에 붙여야 한다.
-             * onDelete를 붙여야 edit button을 누를 때 삭제 버튼이 나타난다.
-             */
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(.plain)
         .navigationTitle("Todo List 📝")
         .toolbar {
             ToolbarItem(placement: .navigationBarLeading) {
